requirements.txt
Flask==2.1.1
python-binance==1.1.1
python-dotenv==0.19.2
slack-bolt==1.13.2
slack-sdk==3.13.1
websocket-client==1.3.1
requests==2.27.1